+++ storage/smartengine/core/db/db_impl.h
@@ -276,14 +276,14 @@ class DBImpl : public DB {
 
   virtual int create_backup_snapshot(BackupSnapshotId backup_id,
                                      MetaSnapshotSet &meta_snapshots,
-                                     int32_t &last_manifest_file_num,
+                                     int64_t &last_manifest_file_num,
                                      uint64_t &last_manifest_file_size,
                                      uint64_t &last_wal_file_num,
                                      BinlogPosition &last_binlog_pos) override;
 
   virtual int record_incremental_extent_ids(const std::string &backup_tmp_dir_path,
-                                            const int32_t first_manifest_file_num,
-                                            const int32_t last_manifest_file_num,
+                                            const int64_t first_manifest_file_num,
+                                            const int64_t last_manifest_file_num,
                                             const uint64_t last_manifest_file_size) override;
 
   virtual int shrink_table_space(int32_t table_space_id) override;
@@ -556,7 +556,7 @@ class DBImpl : public DB {
   ColumnFamilyHandle* DefaultColumnFamily() const override;
 
   // for hot backup
-  virtual int do_manual_checkpoint(int32_t &start_manifest_file_num) override;
+  virtual int do_manual_checkpoint(int64_t &start_manifest_file_num) override;
 
   // const SnapshotList& snapshots() const { return snapshots_; }
   uint64_t snapshots_count() const;
